moved tests to new user flight list component
